requirements.txt
Click==7.0
cv==1.0.0
cycler==0.10.0
Django==2.2.7
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
fonttools==4.0.2
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.1
pprint==0.1
pyparsing==2.4.5
PyQt5==5.13.2
PyQt5-sip==12.7.0
PySide2==5.13.2
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.32
shiboken2==5.13.2
six==1.13.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
Werkzeug==0.15.2
